codebase-chat/requirements.txt
fastapi
uvicorn
faiss-cpu
google-generativeai

numpy
python-dotenv
requests
pydantic
rank-bm25

nbformat
python-multipart
cohere
tqdm
